+++ src/jwsoup/text/utils.py
@@ -6,4 +6,6 @@ def clean_text(text: str) -> str:
     text = re.sub(r"\+\s*\.", ".", text)
     text = re.sub(r"\*\s*\+\s*;", ";", text)
     text = re.sub(r"\*\s*\+", "", text)
+    text = text.replace(" + ", " ")
+    text = text.replace(" * ", " ")
     return text